+++ core/src/test/scala/org/apache/spark/rdd/PairRDDFunctionsSuite.scala
@@ -47,20 +47,6 @@ class PairRDDFunctionsSuite extends FunSuite with SharedSparkContext {
     assert(valuesFor5.toList.sorted === List(1, 3))
   }
 
-  test("topByKey") {
-    val pairs = sc.parallelize(Array((1, 1), (1, 2), (3, 2), (3, 7), (3, 5), (5, 1), (5, 3)), 2)
-
-    val topMap = pairs.topByKey(2).collectAsMap()
-
-    assert(topMap.size === 3)
-    val valuesFor1 = topMap(1)
-    assert(valuesFor1 === Array(2, 1))
-    val valuesFor3 = topMap(3)
-    assert(valuesFor3 === Array(7, 5))
-    val valuesFor5 = topMap(5)
-    assert(valuesFor5 === Array(3, 1))
-  }
-
   test("groupByKey") {
     val pairs = sc.parallelize(Array((1, 1), (1, 2), (1, 3), (2, 1)))
     val groups = pairs.groupByKey().collect()
